Updated SSE2 version of SSe2BgrToBgra

+++ SimdSse2BgrToBgra_amd64.go
@@ -3,8 +3,9 @@ package gocvsimd
 import "unsafe"
 
 //go:noescape
-func _SimdSse2Bgr48pToBgra32(blue unsafe.Pointer, blueStride int, width int, height int, green unsafe.Pointer, greenStride int, red unsafe.Pointer, redStride int, bgra unsafe.Pointer, bgraStride int, alpha uint8)
+func _SimdSse2Bgr48pToBgra32(blue unsafe.Pointer, blueStride int, width int, height int, green unsafe.Pointer, greenStride int, red unsafe.Pointer, redStride int, bgra unsafe.Pointer, bgraStride int, alpha uint64/*uint8*/)
 
-func SimdSse2Bgr48pToBgra32(blue, green, red, bgra View, alpha uint8) {
+func SimdSse2Bgr48pToBgra32(blue, green, red, bgra View, alpha uint64/*uint8*/) {
+
 	_SimdSse2Bgr48pToBgra32(blue.GetData(), blue.GetStride(), blue.GetWidth(), blue.GetHeight(), green.GetData(), green.GetStride(), red.GetData(), red.GetStride(), bgra.GetData(), bgra.GetStride(), alpha)
 }
